+++ src/main/java/stroom/dashboard/expression/v1/FunctionFactory.java
@@ -47,6 +47,7 @@ public class FunctionFactory {
         add(CeilingSecond.class, CeilingSecond.NAME);
 
         add(ParseDate.class, ParseDate.NAME);
+        add(FormatDate.class, FormatDate.NAME);
 
         add(ExtractAuthorityFromUri.class, ExtractAuthorityFromUri.NAME);
         add(ExtractFragmentFromUri.class, ExtractFragmentFromUri.NAME);